refactor and renamed non-blocking-reference to concurrent resource
